Ajustando delete lamina

+++ src/main/java/com/example/squad2_suporte/service/LaminaService.java
@@ -2,6 +2,7 @@ package com.example.squad2_suporte.service;
 
 import com.example.squad2_suporte.Amostras.mapper.LaminaMapper;
 import com.example.squad2_suporte.Lamina.Lamina;
+import com.example.squad2_suporte.config.exceptions.RequisicaoInvalidaException;
 import com.example.squad2_suporte.dto.lamina.LaminaDto;
 import com.example.squad2_suporte.dto.lamina.RetornoLaminaDto;
 import com.example.squad2_suporte.repositorios.LaminaRepository;
@@ -29,11 +30,16 @@ public class LaminaService {
     @Transactional
     public void deletarLamina(Long protocolo) {
         var lamina = laminaRepository.findByProtocolo(protocolo);
-        if (lamina == null){
+        if (lamina == null) {
             throw new RuntimeException("Protocolo inválido ou inexistente");
-        } else {
-            laminaRepository.delete(lamina);
         }
+
+        if (lamina.getLote() != null){
+            throw new RequisicaoInvalidaException("A lamina com protocolo " + protocolo + " está associada a um lote e não pode ser deletada.");
+        }
+
+            laminaRepository.delete(lamina);
+
     }
 
     public Lamina buscarLamina(Long protocolo){
